add ref code again and fix prompt
